Fixed: The active user was not properly set after logging in
Previously, the log in controller could set the activeUser value to two types of objects: a User or a JSON dictionary response. This would result in a method not found when the system tried to call the "pk" method on the dictionary instance.

This commit makes it so that the login controller sets the active user to the correct object in all instances.

Fixes #168