New: Glyph is a WLRemoteObject.
This affects the transformers a lot: symbolCollectionTransformer is deleted, because symbolCollections never should have been a remote property anyway.  GlyphsTransformer is now being used again and GameraGlyphs has a 'glyphs' remoteProperty, while Glyphs themselves are WLRemoteObjects.  (So they'll still get clobbered if you sync the parent remoteObject, but that's not a problem because we will not have to do that when atomic glyph writes are implemented.)  It's not done yet, but I needed to commit.

This also affects the models a lot (GameraGlyphs, Classifier, PageGlyphs.)

New: server side API for /glyph/uuid.  You can Get, Post, Patch, and Delete individual glyphs.  I'm not sure if the responses are right yet.  A lot of the functions from GameraXML.py are now in Glyph.py.  The server sorts the glyphs on every write to XML.

New: MinimalClassifier.  Added so that the real classifier doesn't have to fill two roles: GETs from /classifiers/ and gets from /classifier/uuid.  The former was confusing GlyphsTransformer (glyphs was null).

Maintenance: Refactored all GETs to /classifiers to use the common _updateClassifierArrayControllerWithResponse and to use MinimalClassifier.

Maintenance: Changed the structure of 'features' from an Array to a classed object.