Fix: The horribly large cell sizing of the collection view.
Used setVerticalMargin, setMinItemSize, setMaxItemSize to make the collectionView work.  I think that it's a bug that the collectionView allows itself to be larger than its superview… but I can make it work by setting Min and Max itemSize to control the itemSize and make the collection view not exhibit that bug.