Fixed: authentication controller falsely recognizing different login
If a browser is logged into the site using session (e.g. via the admin link), it will respond to "auth/status" as logged in, but will still get 401s against other requests that are "token"-based.

This fix checks that the authentication response corresponds with the authentication type and notifies the user if.

This should also be addressed on the server side (for later).