Maintenance: removed TNKit
I didn't see any advantage to the TNToolbar.  I wanted to use it to add toolbar items on-the-fly, but its implementation no longer coincides with CPToolbar (from what I could tell).   Also, removing it seems to get the toolbar working correctly in Chrome (even though the origin of the issue is a Chrome issue).